Every project is better with red/black trees
Implemented the worker thread, mostly.  It uses poll now instead
of select, so I need to build an array of struct pollfd for a
arbitrary user-supplied list of alarms, so what better way to
eliminate duplicates and look up which fd goes to which poll struct
than a rbhash :-)

I also divided up the main .xs file into _action.c and _util.c
and organized the code a lot.