Remove the 'repeat' action
This action is fairly problematic in that it could go into a
fast infinite loop and eat up a lot of resources.  If the loop
didn't succeed on the first try, it's unlikely to improve on
subsequent loops.  The user can just write as many iterations
as they want in the action list.