Complete implementation of the watch_thread.  Tests forthcoming.
This finally resolves all the logical tangles I can see related
to action sequences sleeping, unavailability of POLLRDHUP, and
so on.  I don't have a good solution for absence of POLLRDHUP,
which seems not to be available on OpenBSD or MacOS.  In the usual
cases it probably won't matter because HTTP doesn't send readable
data when closing the connection early.