Overhaul the listenable events, and add missing attributes
The listenable events are now
  - EVENT_SHUT
  - EVENT_EOF
  - EVENT_CLOSE
  - EVENT_IN
  - EVENT_PRI

and Windows, OpenBSD, and Mac default to EVENT_SHUT|EVENT_EOF vs.
the EVENT_SHUT that FreeBSD and Linux use.  This way the flags are
just about exactly representative of what is actually being
listened for.

Then I added a bunch of missing attributes to the Alarm objects,
especially the read-only arrayref of actions.

Also lots of documentation updates, and enhanced some unit tests.