Improved bugfix for blocking signals in watcher thread
Instead of blocking the signals in the thread itself, which subjects
us to a very very tiny race condition, block all signals before
creating the thread and then unblock afterward.  This way the thread
inherits the blocked signals and there is zero chance that the kernel
will decide to run the signal handler inside the new thread.