Fix race condition in t/20-detect-event-eof
This always worked on Linux, but showed up on BSD because the
connect action would run asynchronously and not be available
to accept by the very next line of code.  I realized that while
the connect() needed to be nonblocking, the accept() should be
blocking, and this change solves the problem.